refactor(auth): JWT 구조 개선 및 매퍼/서비스 정리
- JwtUtil은 서명/파싱/검증 전담으로 단순화, AuthTokenService에서 토큰 생성 책임 분리
- AuthService.refresh() createAccessToken 인자 순서(role/userName) 버그 수정 및 테스트 정리
- UserMapper 불필요 메서드 제거 및 gpa 타입을 BigDecimal로 정리하여 DB DECIMAL(3,2)와 일치

+++ apps/backend/src/main/java/com/solsol/heycalendar/domain/User.java
@@ -1,5 +1,6 @@
 package com.solsol.heycalendar.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -20,7 +21,7 @@ public class User {
 	private String userName;
 	private Status status;
 	private int grade;
-	private double gpa;
+	private BigDecimal gpa;
 	private LocalDateTime createdAt;
 	private LocalDateTime updatedAt;
 	private Role role;
@@ -28,5 +29,4 @@ public class User {
 	private Long deptNm;
 	private Long collegeNm;
 	private Long univNm;
-
 }
